+++ sdk-java/src/main/java/ly/count/sdk/java/internal/Transport.java
@@ -35,7 +35,6 @@ import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.TrustManagerFactory;
 import javax.net.ssl.X509TrustManager;
-import ly.count.sdk.java.User;
 import org.json.JSONObject;
 
 /**
@@ -117,11 +116,10 @@ public class Transport implements X509TrustManager {
      * set SSL context, calculate and add checksum, load and send user picture if needed.
      *
      * @param request request to send
-     * @param user user to check for picture
      * @return connection, not {@link HttpURLConnection} yet
      * @throws IOException from {@link HttpURLConnection} in case of error
      */
-    HttpURLConnection connection(final Request request, final User user) throws IOException {
+    HttpURLConnection connection(final Request request) throws IOException {
         String endpoint = request.params.remove(Request.ENDPOINT);
 
         if (!request.params.has("device_id") && config.getDeviceId() != null) {
@@ -134,7 +132,7 @@ public class Transport implements X509TrustManager {
         }
 
         String path = config.getServerURL().toString() + endpoint;
-        String picturePathValue = request.params.remove(UserEditorImpl.PICTURE_PATH);
+        String picturePathValue = request.params.remove(ModuleUserProfile.PICTURE_PATH_KEY);
         boolean usingGET = !config.isHTTPPostForced() && request.isGettable(config.getServerURL()) && Utils.isEmptyOrNull(picturePathValue);
 
         if (!usingGET && !Utils.isEmptyOrNull(picturePathValue)) {
@@ -159,7 +157,7 @@ public class Transport implements X509TrustManager {
             PrintWriter writer = null;
             try {
                 L.d("[network] Picture path value " + picturePathValue);
-                byte[] pictureByteData = picturePathValue == null ? null : getPictureDataFromGivenValue(user, picturePathValue);
+                byte[] pictureByteData = picturePathValue == null ? null : getPictureDataFromGivenValue(picturePathValue);
 
                 if (pictureByteData != null) {
                     String boundary = Long.toHexString(System.currentTimeMillis());
@@ -236,34 +234,22 @@ public class Transport implements X509TrustManager {
 
     /**
      * Returns valid picture information
-     * If we have the bytes, give them
-     * Otherwise load them from disk
+     * Load the picture from disk
      *
-     * @param user
-     * @param picture
-     * @return
+     * @param picturePath path to the picture
+     * @return byte array of the picture
      */
-    byte[] getPictureDataFromGivenValue(User user, String picture) {
-        if (user == null) {
-            return null;
-        }
-
+    byte[] getPictureDataFromGivenValue(String picturePath) {
         byte[] data = null;
-        if (UserEditorImpl.PICTURE_IN_USER_PROFILE.equals(picture)) {
-            //if the value is this special value then we know that we will send over bytes that are already provided by the integrator
-            //those stored bytes are already in a internal data structure, use them
-            data = user.picture();
-        } else {
-            //otherwise we assume it is a local path, and we try to read it from disk
-            try {
-                File file = new File(picture);
-                if (!file.exists()) {
-                    return null;
-                }
-                data = Files.readAllBytes(file.toPath());
-            } catch (Throwable t) {
-                L.w("[Transport] getPictureDataFromGivenValue, Error while reading picture from disk " + t);
+        //we assume it is a local path, and we try to read it from disk
+        try {
+            File file = new File(picturePath);
+            if (!file.exists()) {
+                return null;
             }
+            data = Files.readAllBytes(file.toPath());
+        } catch (Throwable t) {
+            L.w("[Transport] getPictureDataFromGivenValue, Error while reading picture from disk " + t);
         }
 
         return data;
@@ -318,7 +304,7 @@ public class Transport implements X509TrustManager {
                     Class requestOwner = request.owner();
                     request.params.remove(Request.MODULE);
 
-                    connection = connection(request, SDKCore.instance.user());
+                    connection = connection(request);
                     connection.connect();
 
                     int code = connection.getResponseCode();